poem_generator_backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
transformers==4.44.2
Pillow==10.3.0
torch==2.4.1
